+++ mirror/visualisations/core/DeepDream.py
@@ -2,11 +2,8 @@ import torch
 import torchvision.transforms.functional as TF
 
 from torch.autograd import Variable
-from torchvision import transforms
-
 from PIL import Image, ImageFilter, ImageChops
 from .Base import Base
-
 from mirror.visualisations.core.utils import image_net_postprocessing, \
     image_net_preprocessing
